web/backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
attrs==24.2.0
azure-core==1.30.2
azure-identity==1.17.1
azure-storage-blob==12.22.0
beanie==1.26.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
expiringdict==1.2.2
fal_client==0.4.1
fastapi==0.112.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.8
isodate==0.6.1
jiter==0.5.0
joblib==1.4.2
joserfc==1.0.0
launchdarkly-eventsource==1.2.0
launchdarkly-server-sdk==9.7.1
lazy-model==0.2.0
motor==3.5.1
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
numpy==2.1.1
openai==1.51.1
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pymongo==4.8.0
pyRFC3339==1.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
replicate==0.32.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
threadpoolctl==3.5.0
tokenizers==0.20.0
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
workos==5.3.0
yarl==1.9.8
